refactor: one table, one warning format, and the column volumes was hiding
**`volumes` was hiding the most consequential fact about a volume.** The
JSON path has always emitted `External`; the table dropped it. An
external volume is one podup neither creates nor deletes, so a `down -v`
that leaves a volume standing was only explicable by reading the compose
file. There is an EXTERNAL column now.

It also moved onto `ui::Table` from a hand-rolled `{:<40} {:<12}`, which
makes cell escaping and column sizing happen in one place instead of two.
Its private `sanitize_cell` is gone — `Table` does that for every command
now — and its two tests moved to `table.rs`, where the behaviour lives,
plus a third pinning that escaping happens before padding.

**Four sites wrote `podup: warning:` by hand**, bypassing the shared
formatter, so a single `generate quadlet` run emitted the bold-yellow
label from tracing and a plain one from `eprintln!` in the same output.
`warn_orphans` was worse: no prefix and no label at all, so the one
message telling a user their compose file has drifted read as stray
output. All four go through `tracing::warn!` now.

**`ui/mod.rs` reached 389 code lines** against a 300 soft limit, and I
kept adding to it. Its tests moved to `mod_tests.rs` — the same split
`autostart`, the libpod client and `stats` already use. 262 now.

`stats` and `top` deliberately keep their own layouts. `stats` needs
right-aligned numeric columns and `top` prints per-container blocks;
folding them into `Table` would mean teaching the shared abstraction
per-column alignment and a block mode to absorb two cases that genuinely
differ. `volumes` was the dialect that was a duplicate for no reason.

160 integration tests and the output contract pass — the contract asserts
headers *contain* their columns, so adding EXTERNAL is compatible.
